+++ trionyx/trionyx/models.py
@@ -205,15 +205,19 @@ class LogManager(models.BaseManager):
 
     def create_log_entry_by_record(self, record):
         """Create log entry by `logging.LogRecord`"""
+        if record.name == 'celery.app.trace':
+            # Remove celery task id, to make same errors match
+            record.args['id'] = ''
+
         log_hash = hashlib.md5(str(' '.join(str(x) for x in [
             record.pathname,
             record.lineno,
-            record.msg,
+            record.getMessage(),
         ])).encode()).hexdigest()
 
         log, _ = self.get_or_create(log_hash=log_hash, defaults={
             'level': record.levelno,
-            'message': record.msg,
+            'message': record.getMessage(),
             'file_path': record.pathname,
             'file_line': record.lineno,
             'last_event': timezone.now(),