+++ trionyx/__init__.py
@@ -5,4 +5,4 @@ trionyx
 :copyright: 2019 by Maikel Martens
 :license: GPLv3
 """
-__version__ = '2.0.2'
+__version__ = '2.1.0'